requirements.txt
scikit-learn==1.2.2
xgboost==1.7.6
google-cloud-bigquery==3.17.2
google-cloud-storage==2.16.0
pandas==1.5.3
joblib==1.3.2
twilio==8.10.0
apache-airflow-providers-sendgrid==3.6.0
sendgrid==6.11.0